backend/functions/requirements.txt
firebase-admin==6.2.0
firebase-functions==0.1.1
google-cloud-storage==2.10.0
google-cloud-firestore==2.11.1